+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # Release Notes
 
-## [Unreleased](https://github.com/ixocreate/entity-package/compare/0.1.0...develop)
+## [Unreleased](https://github.com/ixocreate/entity-package/compare/0.2.0...develop)
+
+## [v0.2.0 (2019-05-06)](https://github.com/ixocreate/entity-package/compare/0.1.0...0.2.0)
+### Changed
+- Moved Type specific classes to Schema Package
+- Upgrade to Application v0.2
+- Upgrade to Schema Package v0.2
+- Drop Type Package
 
 ## [v0.1.0 (2019-04-19)](https://github.com/ixocreate/entity-package/compare/master...0.1.0)
-
 ### Changed
 - Consolidate Package